add PlayerController
