+++ app/src/main/java/com/example/android/playlistmaker/creator/Creator.kt
@@ -7,6 +7,7 @@ import com.example.android.playlistmaker.player.data.player.AndroidStandardPlaye
 import com.example.android.playlistmaker.player.domain.api.AudioPlayerInteractor
 import com.example.android.playlistmaker.player.domain.api.AudioPlayerRepository
 import com.example.android.playlistmaker.player.domain.impl.AudioPlayerInteractorImpl
+import com.example.android.playlistmaker.player.presentation.PlayerController
 import com.example.android.playlistmaker.search.data.SharedPreferncesRepositoryImpl
 import com.example.android.playlistmaker.search.data.TracksRepositoryImpl
 import com.example.android.playlistmaker.search.data.network.RetrofitNetworkClient
@@ -55,5 +56,9 @@ object Creator {
 
     fun provideSettingsControler(activity: Activity): SettingsController {
         return SettingsController(activity)
+    }
+
+    fun providePlayerController(activity: Activity): PlayerController {
+        return PlayerController(activity)
     }
 }
