+++ app/src/main/java/com/example/android/playlistmaker/settings/ui/SettingsActivity.kt
@@ -2,11 +2,12 @@ package com.example.android.playlistmaker.settings.ui
 
 import android.app.Activity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import com.example.android.playlistmaker.R
 import com.example.android.playlistmaker.creator.Creator
 
 
-class SettingsActivity : Activity() {
+class SettingsActivity : AppCompatActivity() {
 
     private val settingsController = Creator.provideSettingsControler(this)
     override fun onCreate(savedInstanceState: Bundle?) {
